refactor(FNP-55): better imports, delete index files, add type imports
