--- packages/client/src/components/atoms/index.ts
@@ -1,5 +1,0 @@
-export * from './Button';
-export * from './Input';
-export * from './Container';
-export * from './Logo';
-export * from './MenuItem';
