--- packages/client/src/components/index.ts
@@ -1,3 +1,0 @@
-export * from './atoms';
-export * from './molecules';
-export * from './organisms';
