+++ packages/client/src/components/molecules/Form/Form.tsx
@@ -1,10 +1,14 @@
 import React from 'react';
 
 import { Text } from '@gravity-ui/uikit';
-import { FormikErrors, FormikHelpers, FormikValues, useFormik } from 'formik';
-import * as Yup from 'yup';
+import type { FormikErrors, FormikHelpers, FormikValues } from 'formik';
+import { useFormik } from 'formik';
+import type * as Yup from 'yup';
 
-import { Button, Container, Input, InputProps } from 'src/components/atoms';
+import { Button } from 'src/components/atoms/Button';
+import { Container } from 'src/components/atoms/Container';
+import type { InputProps } from 'src/components/atoms/Input';
+import { Input } from 'src/components/atoms/Input';
 
 import styles from './Form.module.css';
