--- packages/client/src/components/molecules/index.ts
@@ -1,4 +1,0 @@
-export * from './Form';
-export * from './Copyright';
-export * from './User';
-export * from './EditableText';
