--- packages/client/src/components/organisms/index.ts
@@ -1,3 +1,0 @@
-export * from './App';
-export * from './Page';
-export * from './Header';
