--- packages/client/src/hoc/index.ts
@@ -1,2 +1,0 @@
-export * from './AuthProvider';
-export * from './ProtectedRoute';
