+++ packages/client/src/pages/ErrorBoundaryPage/ErrorBoundaryPage.tsx
@@ -1,4 +1,5 @@
-import { FC, useEffect } from 'react';
+import type { FC } from 'react';
+import { useEffect } from 'react';
 
 import ErrorBoundary from 'src/services/error/error-boundary';
