+++ packages/client/src/pages/ErrorPage/index.ts
