+++ packages/client/src/pages/SignUpPage/SignUpPage.constants.ts
@@ -1,6 +1,6 @@
 import * as Yup from 'yup';
 
-import { InputProps } from 'src/components/atoms';
+import type { InputProps } from 'src/components/atoms/Input';
 
 const phoneRegExp =
   /^((\\+[1-9]{1,4}[ \\-]*)|(\\([0-9]{2,3}\\)[ \\-]*)|([0-9]{2,4})[ \\-]*)*?[0-9]{3,4}?[ \\-]*[0-9]{3,4}?$/;
