+++ packages/client/src/pages/SignUpPage/SignUpPage.tsx
@@ -3,12 +3,14 @@ import React, { useContext, useState } from 'react';
 import { Text } from '@gravity-ui/uikit';
 import { Navigate, useNavigate } from 'react-router-dom';
 
-import { Container, Page } from 'src/components';
-import { Button, Logo } from 'src/components/atoms';
-import { Form } from 'src/components/molecules';
-import { AuthContext } from 'src/hoc';
+import { Button } from 'src/components/atoms/Button';
+import { Container } from 'src/components/atoms/Container';
+import { Logo } from 'src/components/atoms/Logo';
+import { Form } from 'src/components/molecules/Form';
+import { Page } from 'src/components/organisms/Page';
+import { AuthContext } from 'src/hoc/AuthProvider';
 import AuthAPI from 'src/services/api/auth-api';
-import { TSignUpRequest, TUser } from 'src/shared/types/user';
+import type { TSignUpRequest, TUser } from 'src/shared/types/user';
 import { PAGE_ROUTES } from 'src/utils/constants';
 import Helpers from 'src/utils/helpers';
