--- packages/client/src/pages/index.ts
@@ -1,15 +1,0 @@
-export * from './SignInPage';
-export * from './SignUpPage';
-export * from './ErrorPages';
-export * from './ResetPasswordPage';
-export * from './RecoverPasswordPage';
-export * from './LeaderBoardPage';
-export * from './GameResultPage';
-export * from './GameMenuPage';
-export * from './GameSetupPage';
-export * from './GamePage';
-export * from './ProfilePage';
-export * from './GamePausePage';
-export * from './ErrorBoundaryPage';
-export * from './ForumPage';
-export * from './ForumTopicPage';
