+++ packages/client/src/services/error/error-boundary.tsx
@@ -1,4 +1,5 @@
-import React, { Component, ErrorInfo, ReactNode } from 'react';
+import type { ErrorInfo, ReactNode } from 'react';
+import { Component } from 'react';
 
 import Helpers from 'src/utils/helpers';
