+++ packages/client/src/shared/types/global.ts
@@ -1,3 +1,1 @@
 export type Nullable<T> = T | null;
-
-declare let self: ServiceWorkerGlobalScope;
