+++ packages/client/src/utils/service-worker/index.ts
@@ -1,0 +1,1 @@
+export * from './service-worker';
