--- packages/client/src/utils/setup-service-worker/index.ts
@@ -1,1 +1,0 @@
-export * from './setup-service-worker';
